Add yaml wrapper module
